+++ gnome-window-manager/index.js
@@ -3,5 +3,6 @@ module.exports = {
     "gwm-toggle-fullscreen": require("./gwm-toggle-fullscreen"),
     "gwm-almost-maximize": require("./gwm-almost-maximize"),
     "gwm-almost-minimize": require("./gwm-almost-minimize"),
+    "gwm-center": require("./gwm-center"),
   },
 };
