+++ gnome-window-manager/utils.js
@@ -101,6 +101,77 @@ const updateWindowStates = async (exec, windowId, operation, states) => {
   }
 };
 
+const parseKeyValueOutput = (text) =>
+  text
+    .split("\n")
+    .map((line) => line.trim())
+    .filter(Boolean)
+    .reduce((acc, line) => {
+      const [key, value] = line.split("=");
+      if (key && value !== undefined) {
+        acc[key.trim()] = value.trim();
+      }
+      return acc;
+    }, {});
+
+const parseMonitors = (xrandrOutput) =>
+  xrandrOutput
+    .split("\n")
+    .map((line) => line.trim())
+    .map((line) => {
+      const match = line.match(
+        /^([A-Za-z0-9-]+)\s+connected(?:\s+primary)?\s+(\d+)x(\d+)\+(\d+)\+(\d+)/
+      );
+
+      if (!match) {
+        return null;
+      }
+
+      const [, name, width, height, x, y] = match;
+      return {
+        name,
+        width: Number(width),
+        height: Number(height),
+        x: Number(x),
+        y: Number(y),
+      };
+    })
+    .filter(Boolean);
+
+const findMonitorForWindow = (monitors, windowGeometry) => {
+  const windowX = Number(windowGeometry.X);
+  const windowY = Number(windowGeometry.Y);
+  const windowWidth = Number(windowGeometry.WIDTH);
+  const windowHeight = Number(windowGeometry.HEIGHT);
+
+  if ([windowX, windowY, windowWidth, windowHeight].some(Number.isNaN)) {
+    return null;
+  }
+
+  const centerX = windowX + windowWidth / 2;
+  const centerY = windowY + windowHeight / 2;
+
+  return (
+    monitors.find(
+      (monitor) =>
+        centerX >= monitor.x &&
+        centerX <= monitor.x + monitor.width &&
+        centerY >= monitor.y &&
+        centerY <= monitor.y + monitor.height
+    ) || monitors[0] || null
+  );
+};
+
+const delay = (ms) => new Promise((resolve) => setTimeout(resolve, ms));
+
+const getWindowGeometry = async (exec, windowIdDecimal) =>
+  parseKeyValueOutput(
+    await execCommand(exec, `xdotool getwindowgeometry --shell ${windowIdDecimal}`)
+  );
+
+const getMonitors = async (exec) =>
+  parseMonitors(await execCommand(exec, "xrandr --current"));
+
 module.exports = {
   execCommand,
   ensureDependency,
@@ -117,4 +188,10 @@ module.exports = {
   isMaximizedState,
   isFullscreenState,
   updateWindowStates,
+  parseKeyValueOutput,
+  parseMonitors,
+  findMonitorForWindow,
+  delay,
+  getWindowGeometry,
+  getMonitors,
 };
